app/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
mcp
docker
numpy
scikit-learn
fastapi
uvicorn
openai
pydantic
pydantic-settings
python-dotenv
sse-starlette
httpx
rich
playwright>=1.42.0
markdownify
docker
websockets
motor>=3.3.2
pymongo>=4.6.1
beanie>=1.25.0
async-lru>=2.0.0
redis>=5.0.1
beautifulsoup4>=4.12.0
python-multipart
mcp>=1.9.0
pyjwt[crypto]>=2.8.0
cryptography>=3.4.8